v3

+++ ip_parser/ip_parser.h
@@ -4,14 +4,15 @@
 #include <iostream>
 #include <fstream>
 
+//using ip_t = std::tuple<int, int, int, int>;
 using ip_t = std::tuple<unsigned char, unsigned char, unsigned char, unsigned char>;
 
 static std::ostream& operator<<(std::ostream& os, const ip_t& ip)
 {
-	os << std::get<0>(ip)
-		<< "." << std::get<1>(ip)
-		<< "." << std::get<2>(ip)
-		<< "." << std::get<3>(ip);
+	os << (int)std::get<0>(ip)
+		<< "." << (int)std::get<1>(ip)
+		<< "." << (int)std::get<2>(ip)
+		<< "." << (int)std::get<3>(ip);
 	return os;
 }
